EXT-992: getContext() (#44)
Implemented getContext().

By sending a getContext message (no parameters needed) to the container, the container will respond with a get-context message. This message contains the story from the visual editor in the story property.

An alternative to getContext is to use loaded. This will also result in a message containing the current story (amongst other information).

To allow us to verify this functionality in the UI, this pull request contains an example integration between the container and demo apps. The container has a button that allow you to mutate a property on the story content. The field plugin demo prints the story that was received most recently, and has a button to request the context story.