EXT-1159: Make createPluginActions Pure (#60)
createPluginActions is made pure.

createPluginActions does no longer call createPluginMessageListener() directly, instead this responsibility is passed over to createFieldPlugin.

createPluginMessageListener() registers event listener on the window that listens for postMessage messages from the container. createPluginActions creates an object (messageCallbacks) that contains all the callback functions that are to be called for each corresponding plugin message event. Instead of calling createPluginMessageListener with this argument directly, that object is returned so that createFieldPlugin can call createPluginMessageListener instead.

Since the function is now pure, it does not need to run any cleanup.